Move test data to appropriate directory and clean up commented code
