+++ src/apiCalls.js
@@ -1,5 +1,3 @@
-// Your fetch requests will live here!
-
 function fetchRecipesData() {
   return fetch(`https://what-s-cookin-starter-kit.herokuapp.com/api/v1/recipes`)
     .then((response) => response.json())
@@ -19,10 +17,6 @@ function fetchUsersData() {
     .then((response) => response.json())
     .catch((error) => console.log(error));
 }
-
-// function fetchAllData() {
-//   return Promise.all([fetchData('recipes'), fetchData('ingredients'), fetchData('users')]);
-// }
 
 export {
   fetchRecipesData,
